pyproject.toml
[project]
name = "tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = ["fh-utils>=0.4.0", "python-fasthtml>=0.6.8"]

[project.scripts]
start_tutorial = "tutorial:start"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = [
    "ipykernel>=6.29.5",
    "lxml>=5.3.0",
    "pandas>=2.2.2",
    "pytest-cov>=5.0.0",
    "pytest-playwright>=0.5.2",
    "pytest>=8.3.2",
    "ruff>=0.6.3",
    # "devicorn",
]

[tool.uv.sources]
# fh-utils = { path = "../fh_utils", editable = true }
# devicorn = { path = "../devicorn", editable = true }

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.pytest.ini_options]
addopts = "--base-url http://0.0.0.0:5001"
